+++ pages/event-types/index.tsx
@@ -1,12 +1,17 @@
 // TODO: replace headlessui with radix-ui
+import { Menu, Transition } from "@headlessui/react";
+import { UsersIcon } from "@heroicons/react/solid";
 import { ChevronDownIcon, PlusIcon } from "@heroicons/react/solid";
+import { DotsHorizontalIcon, ExternalLinkIcon, LinkIcon } from "@heroicons/react/solid";
 import { SchedulingType } from "@prisma/client";
 import Head from "next/head";
+import Link from "next/link";
 import { useRouter } from "next/router";
 import React, { Fragment, useRef } from "react";
 import { useMutation } from "react-query";
 
 import { QueryCell } from "@lib/QueryCell";
+import classNames from "@lib/classNames";
 import { HttpError } from "@lib/core/http/error";
 import { useLocale } from "@lib/hooks/useLocale";
 import { useToggleQuery } from "@lib/hooks/useToggleQuery";
@@ -17,10 +22,12 @@ import { CreateEventType } from "@lib/types/event-type";
 
 import { Dialog, DialogClose, DialogContent } from "@components/Dialog";
 import Shell from "@components/Shell";
-import EventTypeList from "@components/pages/eventtypes/EventTypeList";
-import EventTypeListHeading from "@components/pages/eventtypes/EventTypeListHeading";
+import { Tooltip } from "@components/Tooltip";
+import EventTypeDescription from "@components/eventtype/EventTypeDescription";
 import { Alert } from "@components/ui/Alert";
 import Avatar from "@components/ui/Avatar";
+import AvatarGroup from "@components/ui/AvatarGroup";
+import Badge from "@components/ui/Badge";
 import { Button } from "@components/ui/Button";
 import Dropdown, {
   DropdownMenuContent,
@@ -54,6 +61,211 @@ const CreateFirstEventTypeView = ({ canAddEvents, profiles }: CreateEventTypePro
   );
 };
 
+type EventTypeGroup = inferQueryOutput<"viewer.eventTypes">["eventTypeGroups"][number];
+type EventType = EventTypeGroup["eventTypes"][number];
+interface EventTypeListProps {
+  profile: { slug: string | null };
+  readOnly: boolean;
+  types: EventType[];
+}
+const EventTypeList = ({ readOnly, types, profile }: EventTypeListProps): JSX.Element => {
+  const { t } = useLocale();
+  return (
+    <div className="mb-16 -mx-4 overflow-hidden bg-white border border-gray-200 rounded-sm sm:mx-0">
+      <ul className="divide-y divide-neutral-200" data-testid="event-types">
+        {types.map((type) => (
+          <li
+            key={type.id}
+            className={classNames(
+              type.$disabled && "opacity-30 cursor-not-allowed pointer-events-none select-none"
+            )}
+            data-disabled={type.$disabled ? 1 : 0}>
+            <div
+              className={classNames(
+                "hover:bg-neutral-50 flex justify-between items-center ",
+                type.$disabled && "pointer-events-none"
+              )}>
+              <div className="flex items-center justify-between w-full px-4 py-4 sm:px-6 hover:bg-neutral-50">
+                <Link href={"/event-types/" + type.id}>
+                  <a
+                    className="flex-grow text-sm truncate"
+                    title={`${type.title} ${type.description ? `– ${type.description}` : ""}`}>
+                    <div>
+                      <span className="font-medium truncate text-neutral-900">{type.title}</span>
+                      {type.hidden && (
+                        <span className="ml-2 inline items-center px-1.5 py-0.5 rounded-sm text-xs font-medium bg-yellow-100 text-yellow-800">
+                          {t("hidden")}
+                        </span>
+                      )}
+                      {readOnly && (
+                        <span className="ml-2 inline items-center px-1.5 py-0.5 rounded-sm text-xs font-medium bg-gray-100 text-gray-800">
+                          {t("readonly")}
+                        </span>
+                      )}
+                    </div>
+                    <EventTypeDescription eventType={type} />
+                  </a>
+                </Link>
+
+                <div className="flex-shrink-0 hidden mt-4 sm:flex sm:mt-0 sm:ml-5">
+                  <div className="flex items-center space-x-2 overflow-hidden">
+                    {type.users?.length > 1 && (
+                      <AvatarGroup
+                        size={8}
+                        truncateAfter={4}
+                        items={type.users.map((organizer) => ({
+                          alt: organizer.name || "",
+                          image: organizer.avatar || "",
+                        }))}
+                      />
+                    )}
+                    <Tooltip content={t("preview")}>
+                      <a
+                        href={`${process.env.NEXT_PUBLIC_APP_URL}/${profile.slug}/${type.slug}`}
+                        target="_blank"
+                        rel="noreferrer"
+                        className="btn-icon">
+                        <ExternalLinkIcon className="w-5 h-5 group-hover:text-black" />
+                      </a>
+                    </Tooltip>
+
+                    <Tooltip content={t("copy_link")}>
+                      <button
+                        onClick={() => {
+                          showToast(t("link_copied"), "success");
+                          navigator.clipboard.writeText(
+                            `${process.env.NEXT_PUBLIC_APP_URL}/${profile.slug}/${type.slug}`
+                          );
+                        }}
+                        className="btn-icon">
+                        <LinkIcon className="w-5 h-5 group-hover:text-black" />
+                      </button>
+                    </Tooltip>
+                  </div>
+                </div>
+              </div>
+              <div className="flex flex-shrink-0 mr-5 sm:hidden">
+                <Menu as="div" className="inline-block text-left">
+                  {({ open }) => (
+                    <>
+                      <div>
+                        <Menu.Button className="p-2 mt-1 border border-transparent text-neutral-400 hover:border-gray-200">
+                          <span className="sr-only">{t("open_options")}</span>
+                          <DotsHorizontalIcon className="w-5 h-5" aria-hidden="true" />
+                        </Menu.Button>
+                      </div>
+
+                      <Transition
+                        show={open}
+                        as={Fragment}
+                        enter="transition ease-out duration-100"
+                        enterFrom="transform opacity-0 scale-95"
+                        enterTo="transform opacity-100 scale-100"
+                        leave="transition ease-in duration-75"
+                        leaveFrom="transform opacity-100 scale-100"
+                        leaveTo="transform opacity-0 scale-95">
+                        <Menu.Items
+                          static
+                          className="absolute right-0 z-10 w-56 mt-2 origin-top-right bg-white divide-y rounded-sm shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none divide-neutral-100">
+                          <div className="py-1">
+                            <Menu.Item>
+                              {({ active }) => (
+                                <a
+                                  href={`${process.env.NEXT_PUBLIC_APP_URL}/${profile.slug}/${type.slug}`}
+                                  target="_blank"
+                                  rel="noreferrer"
+                                  className={classNames(
+                                    active ? "bg-neutral-100 text-neutral-900" : "text-neutral-700",
+                                    "group flex items-center px-4 py-2 text-sm font-medium"
+                                  )}>
+                                  <ExternalLinkIcon
+                                    className="w-4 h-4 mr-3 text-neutral-400 group-hover:text-neutral-500"
+                                    aria-hidden="true"
+                                  />
+                                  {t("preview")}
+                                </a>
+                              )}
+                            </Menu.Item>
+                            <Menu.Item>
+                              {({ active }) => (
+                                <button
+                                  onClick={() => {
+                                    showToast("Link copied!", "success");
+                                    navigator.clipboard.writeText(
+                                      `${process.env.NEXT_PUBLIC_APP_URL}/${profile.slug}/${type.slug}`
+                                    );
+                                  }}
+                                  className={classNames(
+                                    active ? "bg-neutral-100 text-neutral-900" : "text-neutral-700",
+                                    "group flex items-center px-4 py-2 text-sm w-full font-medium"
+                                  )}>
+                                  <LinkIcon
+                                    className="w-4 h-4 mr-3 text-neutral-400 group-hover:text-neutral-500"
+                                    aria-hidden="true"
+                                  />
+                                  {t("copy_link")}
+                                </button>
+                              )}
+                            </Menu.Item>
+                          </div>
+                        </Menu.Items>
+                      </Transition>
+                    </>
+                  )}
+                </Menu>
+              </div>
+            </div>
+          </li>
+        ))}
+      </ul>
+    </div>
+  );
+};
+
+interface EventTypeListHeadingProps {
+  profile: Profile;
+  membershipCount: number;
+}
+const EventTypeListHeading = ({ profile, membershipCount }: EventTypeListHeadingProps): JSX.Element => (
+  <div className="flex mb-4">
+    <Link href="/settings/teams">
+      <a>
+        <Avatar
+          alt={profile?.name || ""}
+          imageSrc={profile?.image || undefined}
+          size={8}
+          className="inline mt-1 mr-2"
+        />
+      </a>
+    </Link>
+    <div>
+      <Link href="/settings/teams">
+        <a className="font-bold">{profile?.name || ""}</a>
+      </Link>
+      {membershipCount && (
+        <span className="relative ml-2 text-xs text-neutral-500 -top-px">
+          <Link href="/settings/teams">
+            <a>
+              <Badge variant="gray">
+                <UsersIcon className="inline w-3 h-3 mr-1 -mt-px" />
+                {membershipCount}
+              </Badge>
+            </a>
+          </Link>
+        </span>
+      )}
+      {profile?.slug && (
+        <Link href={`${process.env.NEXT_PUBLIC_APP_URL}/${profile.slug}`}>
+          <a className="block text-xs text-neutral-500">{`${process.env.NEXT_PUBLIC_APP_URL?.replace(
+            "https://",
+            ""
+          )}/${profile.slug}`}</a>
+        </Link>
+      )}
+    </div>
+  </div>
+);
+
 const EventTypesPage = () => {
   const { t } = useLocale();
   const query = trpc.useQuery(["viewer.eventTypes"]);
@@ -69,7 +281,7 @@ const EventTypesPage = () => {
         subtitle={t("event_types_page_subtitle")}
         CTA={
           query.data &&
-          query.data.eventTypes.length !== 0 && (
+          query.data.eventTypeGroups.length !== 0 && (
             <CreateNewEventButton canAddEvents={query.data.canAddEvents} profiles={query.data.profiles} />
           )
         }>
@@ -92,11 +304,11 @@ const EventTypesPage = () => {
                   className="my-4"
                 />
               )}
-              {data.eventTypes &&
-                data.eventTypes.map((input) => (
+              {data.eventTypeGroups &&
+                data.eventTypeGroups.map((input) => (
                   <Fragment key={input.profile.slug}>
                     {/* hide list heading when there is only one (current user) */}
-                    {(data.eventTypes.length !== 1 || input.teamId) && (
+                    {(data.eventTypeGroups.length !== 1 || input.teamId) && (
                       <EventTypeListHeading
                         profile={input.profile}
                         membershipCount={input.metadata.membershipCount}
@@ -110,7 +322,7 @@ const EventTypesPage = () => {
                   </Fragment>
                 ))}
 
-              {data.eventTypes.length === 0 && (
+              {data.eventTypeGroups.length === 0 && (
                 <CreateFirstEventTypeView profiles={data.profiles} canAddEvents={data.canAddEvents} />
               )}
             </>
